pyproject.toml
[project]
name = "todo-drf"
version = "0.1.0"
description = "todo-drf"
authors = [
    {name = "shmaizer"}
]
packages = [
    { include = "todoapi" },
    { include = "apps" }
]
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "djangorestframework (>=3.16.1,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "drf-yasg (>=1.21.10,<2.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "pre-commit (>=4.3.0,<5.0.0)",
    "django-extensions (>=4.1,<5.0)",
    "werkzeug (>=3.1.3,<4.0.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "black (>=25.9.0,<26.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "pyupgrade (>=3.20.0,<4.0.0)",
    "autoflake (>=2.3.1,<3.0.0)",
    "check-jsonschema (>=0.34.0,<0.35.0)",
]
